+++ projects/batfish-common-protocol/src/main/java/org/batfish/common/plugin/IBatfish.java
@@ -19,7 +19,7 @@ import org.batfish.datamodel.Flow;
 import org.batfish.datamodel.FlowHistory;
 import org.batfish.datamodel.Ip;
 import org.batfish.datamodel.Topology;
-import org.batfish.datamodel.answers.AclLinesAnswerElement;
+import org.batfish.datamodel.answers.AclLinesAnswerElementInterface;
 import org.batfish.datamodel.answers.AnswerElement;
 import org.batfish.datamodel.answers.ConvertConfigurationAnswerElement;
 import org.batfish.datamodel.answers.DataPlaneAnswerElement;
@@ -47,7 +47,7 @@ public interface IBatfish extends IPluginConsumer {
   void answerAclReachability(
       String aclNameRegexStr,
       NamedStructureEquivalenceSets<?> aclEqSets,
-      AclLinesAnswerElement emptyAnswer);
+      AclLinesAnswerElementInterface emptyAnswer);
 
   void checkDataPlane();
